+++ project/OsEngine/Market/Servers/BloFin/BloFinFutures/BloFinFuturesServer.cs
@@ -1219,8 +1219,8 @@ namespace OsEngine.Market.Servers.BloFin
 
                 for (int i = 0; i < responseDepth.data.asks.Count; i++)
                 {
-                    decimal ask = responseDepth.data.asks[i][1].ToString().ToDecimal();
-                    decimal price = responseDepth.data.asks[i][0].ToString().ToDecimal();
+                    double ask = responseDepth.data.asks[i][1].ToString().ToDouble();
+                    double price = responseDepth.data.asks[i][0].ToString().ToDouble();
 
                     if (ask == 0 ||
                         price == 0)
@@ -1236,8 +1236,8 @@ namespace OsEngine.Market.Servers.BloFin
 
                 for (int i = 0; i < responseDepth.data.bids.Count; i++)
                 {
-                    decimal bid = responseDepth.data.bids[i][1].ToString().ToDecimal();
-                    decimal price = responseDepth.data.bids[i][0].ToString().ToDecimal();
+                    double bid = responseDepth.data.bids[i][1].ToString().ToDouble();
+                    double price = responseDepth.data.bids[i][0].ToString().ToDouble();
 
                     if (bid == 0 ||
                         price == 0)
